+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 ### General
 
 - improve documentation of how Flask Unchained works
-- updated to py-meta-utils 0.7.3 and flask-sqlalchemy-unchained 0.6.6
+- updated to py-meta-utils 0.7.3 and flask-sqlalchemy-unchained 0.6.7
 - update to marshmallow 2.16
 
 ### Breaking Changes